Moves nearby screen to root screen folder, removing nested navigation stack
